feat: v1.8.0 — full-surface API exposure, fixed assets, revenue recognition
- ~140 new REST endpoints: purchase orders/suppliers, general ledger
  (journals, trial balance, balance sheet, income statement, periods),
  AP (bills, payment runs, aging, 3-way match), AR (aging, payment
  application, credit memos, write-offs, dunning, statements), warehouse
  (locations, adjust/move, cycle counts), fulfillment (waves, pick/pack/
  ship), receiving (receipts, put-away), work orders, quality, BOM,
  lots, serials — all OpenAPI-documented
- New modules (full stack): fixed-asset register with depreciation
  schedules and revenue recognition (ASC 606) with recognition schedules
- Cycle-count workflow with transactional variance application
- Config-gated GL auto-posting for depreciation and revenue recognition
- Purchase-order state machine enforced at the store layer
- PostgreSQL parity for nine formerly SQLite-only stores
- ~290 new tests; model-test backfill for six untested models
- Fixes: lot-less location-inventory NOT NULL failure;
  create_receipt_from_po nonexistent-column query
- Zero clippy warnings workspace-wide

+++ crates/stateset-db/src/postgres/custom_objects.rs
@@ -239,7 +239,9 @@ impl PgCustomObjectRepository {
         };
 
         let now = Utc::now();
-        let updated: (i64,) = sqlx::query_as(
+        // `RETURNING 1` is an INT4 literal, so decode it as i32 (an i64 decode
+        // fails with a type mismatch and was breaking the update on Postgres).
+        let updated: (i32,) = sqlx::query_as(
             "UPDATE custom_object_types
              SET display_name = $1, description = $2, fields = $3, updated_at = $4, version = version + 1
              WHERE id = $5 AND version = $6
